Move `FeedImageDataLoaderPresentationAdapter` to it's own file
